examples/langgraph-python/pyproject.toml
[project]
name = "langgraph-python"
version = "0.1.0"
description = "LangGraph Python example with code interpreter"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "agbcloud-sdk",
    "python-dotenv",
    "langchain-openai",
    "langchain-core",
    "langgraph",
]
